Extract shaders to separate files
