+++ src/renderer/components/visualizers/SpectrumAnalyzer.vue
@@ -4,6 +4,8 @@ import ShaderCanvas from "@/components/ShaderCanvas.vue";
 import {VISUALIZER_BIN_COUNT} from "@shared/constants";
 import * as THREE from "three";
 import { watch } from "vue";
+import {SpectrogramShader} from "@/shaders/components/SpectrogramShader";
+import {SpectrumShader} from "@/shaders/components/SpectrumShader";
 
 const props = defineProps<{
   node: AudioNode,
@@ -48,52 +50,13 @@ const uniformData = {
   u_amplitudes: {value: new Float32Array(VISUALIZER_BIN_COUNT)},
 };
 
-const spectrumShader = `
-  precision highp float;
-
-  uniform vec2 u_resolution;
-  uniform float[${VISUALIZER_BIN_COUNT}] u_amplitudes;
-  uniform vec3 u_color;
-
-  void main(){
-    vec2 uv = gl_FragCoord.xy / u_resolution;
-    float amplitude = u_amplitudes[int(uv.x * float(u_amplitudes.length()))];
-    float underCurve = 0.0;
-    if(uv.y <= amplitude) {
-      underCurve = 1.0;
-    }
-    gl_FragColor = vec4(u_color * underCurve, underCurve);
-  }`;
-
-const spectrogramShader = `
-  precision highp float;
-
-  uniform vec2 u_resolution;
-  uniform sampler2D u_backbuffer;
-  uniform float[${VISUALIZER_BIN_COUNT}] u_amplitudes;
-  uniform vec3 u_color;
-
-  void main(){
-    vec2 uv = gl_FragCoord.xy / u_resolution;
-    float amplitude = u_amplitudes[int(uv.y * float(u_amplitudes.length()))];
-    amplitude = clamp(pow(amplitude, 2.0), 0.0, 1.0);
-
-    // If first column
-    if (gl_FragCoord.x == 0.5) {
-      gl_FragColor = vec4(u_color - amplitude, 1.0);
-    } else {
-      // Get the color of the previous column (x-1)
-      gl_FragColor = texture2D(u_backbuffer, vec2(uv.x - (1.0 / u_resolution.x), uv.y));
-    }
-  }`;
-
 const render = (uniforms: Record<string, any>) => {
   const spectrum = new Uint8Array(analyser.frequencyBinCount);
   analyser.getByteFrequencyData(spectrum);
   uniforms.u_amplitudes.value = logParabolicSpectrum(spectrum, VISUALIZER_BIN_COUNT);
 };
 
-const getShader = () => props.spectrogram ? spectrogramShader : spectrumShader;
+const getShader = () => props.spectrogram ? SpectrogramShader : SpectrumShader;
 </script>
 
 <template>
